+++ pkg/api/v1beta4/dynakube/otelc_props_test.go
@@ -3,7 +3,7 @@ package dynakube
 import (
 	"testing"
 
-	"github.com/Dynatrace/dynatrace-operator/pkg/api/v1beta3/dynakube/activegate"
+	"github.com/Dynatrace/dynatrace-operator/pkg/api/v1beta4/dynakube/activegate"
 	"github.com/stretchr/testify/assert"
 )
